add support for cells.
adds a new config group (NOVA_CELLS) which enables setup of cells in a
packstack environment. it greatly replaces the usage of
CONFIG_CONTROLLER_HOST with the new CONFIG_NOVA_CELLS_PARENT_HOST and
CONFIG_NOVA_CELLS_CHILD_HOSTS where appropriate.

it allows setting of a single parent (top) cell and a set of child
cells.

for the moment a single rabbitmq host is set.

previously to submit this patch:
- review the usage of rabbitmq_host (might be better to have local cell
  instances)
- review the setup of compute_hosts (specify cells)
- review documentation